requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+git://github.com/anastasia/namecolor.git@ca31a23#egg=namecolor
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask, pip-tools
cloudpickle==0.6.1        # via scikit-image
colormath==3.0.0
cycler==0.10.0            # via matplotlib
dask[array]==1.0.0        # via scikit-image
decorator==4.3.0          # via ipython, networkx, traitlets
flask-sqlalchemy==2.3.2
flask==1.0.2
gunicorn==19.9.0
idna==2.7                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1
itsdangerous==1.1.0       # via flask
jedi==0.13.1              # via ipython
jinja2==2.10              # via flask
kiwisolver==1.0.1         # via matplotlib
markupsafe==1.1.0         # via jinja2
matplotlib==3.0.2         # via scikit-image
networkx==2.2             # via colormath, scikit-image
numpy==1.15.4             # via colormath, dask, matplotlib, pywavelets, scipy
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==5.3.0             # via scikit-image
pip-tools==3.1.0
prompt-toolkit==2.0.7     # via ipython
psycopg2==2.7.6.1
ptyprocess==0.6.0         # via pexpect
pygments==2.3.0           # via ipython
pyparsing==2.3.0          # via matplotlib
python-colourlovers==0.1.1
python-dateutil==2.7.5    # via matplotlib
pywavelets==1.0.1         # via scikit-image
requests==2.20.1          # via python-colourlovers
scikit-image==0.14.1
scipy==1.1.0              # via scikit-image
six==1.11.0               # via cycler, pip-tools, prompt-toolkit, python-dateutil, scikit-image, traitlets
sqlalchemy==1.2.14        # via flask-sqlalchemy
toolz==0.9.0              # via dask
torch==0.4.1
traitlets==4.3.2          # via ipython
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1          # via flask
